requirements_no_hashes.txt
# requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
open3d==0.19.0
scipy==1.13.1
scikit-learn==1.3.1
matplotlib==3.8.4
tqdm==4.66.1
pyyaml==6.0.1
pandas==2.1.1
